refactor: Pull all the parameter handler code out into its own file
Parameters is, in the language of the spec, a Payload Template
